Finished init_wiki. Restructure auth things into one module. Rename auth stuff to be more descriptive.
